# contracts/openapi/operator-api-v1.yaml
openapi: 3.1.0
info:
  title: SRE Operator API
  version: 1.0.0
  description: The sole versioned authenticated REST boundary used by the operator UI.
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/incidents:
    get:
      operationId: listIncidents
      summary: List authorized incidents using cursor pagination.
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/From'
        - $ref: '#/components/parameters/To'
        - name: teamId
          in: query
          required: false
          schema:
            type: string
            format: uuid
        - name: projectId
          in: query
          required: false
          schema:
            type: string
            format: uuid
        - name: environmentId
          in: query
          required: false
          schema:
            type: string
            format: uuid
        - name: serviceId
          in: query
          required: false
          schema:
            type: string
            format: uuid
        - name: severity
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/Severity'
        - name: status
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/IncidentStatus'
        - name: rcaStatus
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/RcaRunStatus'
        - name: assigneeId
          in: query
          required: false
          schema:
            type: string
            format: uuid
        - name: search
          in: query
          required: false
          schema:
            type: string
            minLength: 1
            maxLength: 200
        - name: sortBy
          in: query
          required: false
          schema:
            type: string
            enum: [openedAt, updatedAt, severity, status]
            default: updatedAt
        - name: sortOrder
          in: query
          required: false
          schema:
            type: string
            enum: [asc, desc]
            default: desc
      responses:
        <<: *commonErrorResponses
        '200':
          description: Cursor page of incidents.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CursorPageIncidents'
  /api/v1/incidents/{id}:
    parameters:
      - $ref: '#/components/parameters/ResourceId'
    get:
      operationId: getIncident
      summary: Get an authorized incident.
      responses:
        <<: *commonErrorResponses
        '200':
          description: Incident detail at the returned version.
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IncidentDetail'
...
  /api/v1/alerts/{id}:
    parameters:
      - $ref: '#/components/parameters/ResourceId'
    get:
      operationId: getAlert
      summary: Get an authorized alert.
      responses:
        <<: *commonErrorResponses
        '200':
          description: Alert detail without the immutable raw webhook payload.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlertDetail'
...
  /api/v1/incidents/{id}/rca-runs:
    parameters:
      - $ref: '#/components/parameters/ResourceId'
    get:
      operationId: listIncidentRcaRuns
      summary: List immutable RCA runs for an incident.
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/From'
        - $ref: '#/components/parameters/To'
      responses:
        <<: *commonErrorResponses
        '200':
          description: Cursor page of RCA runs.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CursorPageRcaRuns'
    post:
      operationId: createIncidentRcaRun
      summary: Start or replay an RCA run idempotently.
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRcaRunRequest'
      responses:
        <<: *commonErrorResponses
        '202':
          description: RCA run accepted or replayed from the same idempotency key.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RcaRun'
...
  /api/v1/rca-runs/{id}/report:
    parameters:
      - $ref: '#/components/parameters/ResourceId'
    get:
      operationId: getRcaReport
      summary: Get the immutable report for an RCA run.
      responses:
        <<: *commonErrorResponses
        '200':
          description: Versioned RCA report.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RcaReport'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    ResourceId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    Cursor:
      name: cursor
      in: query
      required: false
      description: Opaque cursor from the previous page; it is bound to the active filters.
      schema:
        type: string
        minLength: 1
        maxLength: 2048
    Limit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 50
    From:
      name: from
      in: query
      required: true
      schema:
        $ref: '#/components/schemas/UtcDateTime'
    To:
      name: to
      in: query
      required: true
      schema:
        $ref: '#/components/schemas/UtcDateTime'
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 1
        maxLength: 128
  headers:
    ETag:
      description: Strong version ETag used by If-Match mutations.
      required: true
      schema:
        type: string
        pattern: '^"[^"]+"$'
...
  schemas:
    UtcDateTime:
      type: string
      format: date-time
      pattern: 'Z$'
      description: UTC timestamp serialized in RFC 3339 Z notation.
    AlertState:
      type: string
      enum: [FIRING, RESOLVED]
...
    IncidentStatus:
      type: string
      enum: [OPEN, INVESTIGATING, RESOLVED]
    RcaRunStatus:
      type: string
      enum:
        - WAITING_FOR_CLASSIFICATION
        - QUEUED
        - RUNNING
        - SUCCEEDED
        - PARTIAL
        - FAILED
        - CANCELLED
    EvidenceRelation:
      type: string
      enum: [SUPPORTS, CONTRADICTS, MISSING]
...
    Severity:
      type: string
      enum: [CRITICAL, HIGH, MEDIUM, LOW]
    Provider:
      type: string
      enum: [GCP, AWS]
    CanonicalSeverity:
      type: string
      enum: [SEV1, SEV3, UNMAPPED]
    ProblemCode:
      type: string
      enum:
        - INVALID_REQUEST
        - INVALID_CURSOR
        - INVALID_TIME_RANGE
        - UNAUTHENTICATED
        - SCOPE_FORBIDDEN
        - RESOURCE_NOT_FOUND
        - INCIDENT_VERSION_CONFLICT
        - INVALID_INCIDENT_TRANSITION
        - RCA_ALREADY_RUNNING
        - MCP_TIMEOUT
        - INTERNAL_ERROR
...
    Scope:
      type: object
      required: [teamId, projectId, environmentId, serviceId]
      properties:
        teamId:
          type: [string, 'null']
          format: uuid
        projectId:
          type: [string, 'null']
          format: uuid
        environmentId:
          type: [string, 'null']
          format: uuid
        serviceId:
          type: [string, 'null']
          format: uuid
      additionalProperties: false
    Assignee:
      type: object
      required: [id, displayName]
      properties:
        id:
          type: string
          format: uuid
        displayName:
          type: string
          minLength: 1
      additionalProperties: false
    IncidentSummary:
      type: object
      required: &incidentSummaryRequired
        - id
        - incidentNumber
        - title
        - severity
        - status
        - alertState
        - rcaStatus
        - scope
        - acknowledged
        - openedAt
        - updatedAt
        - version
      properties: &incidentSummaryProperties
        id:
          type: string
          format: uuid
        incidentNumber:
          type: string
          pattern: '^INC-[0-9]+$'
        title:
          type: string
          minLength: 1
        severity:
          $ref: '#/components/schemas/CanonicalSeverity'
        status:
          $ref: '#/components/schemas/IncidentStatus'
        alertState:
          $ref: '#/components/schemas/AlertState'
        rcaStatus:
          oneOf:
            - $ref: '#/components/schemas/RcaRunStatus'
            - type: 'null'
        scope:
          $ref: '#/components/schemas/Scope'
        acknowledged:
          type: boolean
        acknowledgedAt:
          oneOf:
            - $ref: '#/components/schemas/UtcDateTime'
            - type: 'null'
        acknowledgedBy:
          type: [string, 'null']
          format: uuid
        assignee:
          oneOf:
            - $ref: '#/components/schemas/Assignee'
            - type: 'null'
        openedAt:
          $ref: '#/components/schemas/UtcDateTime'
        updatedAt:
          $ref: '#/components/schemas/UtcDateTime'
        resolvedAt:
          oneOf:
            - $ref: '#/components/schemas/UtcDateTime'
            - type: 'null'
        version:
          type: integer
          minimum: 1
      allOf: &incidentStateInvariants
        - if:
            required: [acknowledged]
            properties:
              acknowledged:
                const: true
          then:
            required: [acknowledgedAt, acknowledgedBy]
            properties:
              acknowledgedAt:
                $ref: '#/components/schemas/UtcDateTime'
              acknowledgedBy:
                type: string
                format: uuid
        - if:
            required: [status]
            properties:
              status:
                const: RESOLVED
          then:
            required: [resolvedAt]
            properties:
              resolvedAt:
                $ref: '#/components/schemas/UtcDateTime'
      additionalProperties: false
    IncidentDetail:
      type: object
      required:
        - id
        - incidentNumber
        - title
        - severity
        - status
        - alertState
        - rcaStatus
        - scope
        - acknowledged
        - openedAt
        - updatedAt
        - version
        - alertIds
        - rcaRunIds
      properties:
        <<: *incidentSummaryProperties
        description:
          type: string
        alertIds:
          type: array
          items:
            type: string
            format: uuid
          uniqueItems: true
        rcaRunIds:
          type: array
          items:
            type: string
            format: uuid
          uniqueItems: true
      allOf: *incidentStateInvariants
      additionalProperties: false
...
    AlertDetail:
      type: object
      required:
        - id
        - sourceId
        - fingerprint
        - title
        - severity
        - state
        - classificationStatus
        - startsAt
        - updatedAt
        - labels
        - annotations
        - provider
        - folderCode
        - alertName
        - severityRaw
        - issue
        - normalizationWarnings
      properties:
        <<: *alertSummaryProperties
        provider:
          $ref: '#/components/schemas/Provider'
        folderCode:
          type: [string, 'null']
        alertName:
          type: [string, 'null']
        severityRaw:
          type: [string, 'null']
        severity:
          $ref: '#/components/schemas/CanonicalSeverity'
        issue:
          $ref: '#/components/schemas/AlertIssue'
        normalization:
          $ref: '#/components/schemas/NormalizationInfo'
        normalizationWarnings:
          type: array
          items:
            type: string
        labels:
          type: object
          additionalProperties: true
        annotations:
          type: object
          additionalProperties:
            type: string
        generatorUrl:
          type: string
          format: uri
      additionalProperties: false
    AlertIssue:
      type: object
      required: [rawText, source, contentType, untrusted]
      properties:
        rawText:
          type: string
        source:
          const: grafana.annotations.AlertValues
        contentType:
          const: text/plain
        untrusted:
          const: true
      additionalProperties: false
    NormalizationInfo:
      type: object
      required: [status, ruleId, ruleVersion]
      properties:
        status:
          type: string
          enum: [NORMALIZED, UNCLASSIFIED, VALIDATION_FAILED]
        ruleId:
          type: [string, 'null']
          format: uuid
        ruleVersion:
          type: [integer, 'null']
          minimum: 1
      additionalProperties: false
    RcaRun:
      type: object
      required: [id, incidentId, runNumber, status, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        incidentId:
          type: string
          format: uuid
        runNumber:
          type: integer
          minimum: 1
        status:
          $ref: '#/components/schemas/RcaRunStatus'
        createdAt:
          $ref: '#/components/schemas/UtcDateTime'
        updatedAt:
          $ref: '#/components/schemas/UtcDateTime'
        startedAt:
          oneOf:
            - $ref: '#/components/schemas/UtcDateTime'
            - type: 'null'
        completedAt:
          oneOf:
            - $ref: '#/components/schemas/UtcDateTime'
            - type: 'null'
        failureCode:
          oneOf:
            - $ref: '#/components/schemas/ProblemCode'
            - type: 'null'
        reportId:
          type: [string, 'null']
          format: uuid
      additionalProperties: false
    EvidenceReference:
      type: object
      required: [evidenceId, partitionTimestamp, relation]
      properties:
        evidenceId:
          type: string
          format: uuid
        partitionTimestamp:
          $ref: '#/components/schemas/UtcDateTime'
        relation:
          $ref: '#/components/schemas/EvidenceRelation'
      additionalProperties: false
    ReportClaim:
      type: object
      required: [statement, evidence]
      properties:
        statement:
          type: string
          minLength: 1
        evidence:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/EvidenceReference'
      additionalProperties: false
    RcaReport:
      type: object
      required:
        - id
        - rcaRunId
        - incidentId
        - reportVersion
        - status
        - summary
        - rootCause
        - impact
        - recommendations
        - claims
        - createdAt
      properties:
        id:
          type: string
          format: uuid
        rcaRunId:
          type: string
          format: uuid
        incidentId:
          type: string
          format: uuid
        reportVersion:
          type: integer
          minimum: 1
        status:
          type: string
          enum: [SUCCEEDED, PARTIAL]
        summary:
          type: string
          minLength: 1
        rootCause:
          type: string
          minLength: 1
        impact:
          type: string
          minLength: 1
        recommendations:
          type: array
          items:
            type: string
            minLength: 1
        claims:
          type: array
          items:
            $ref: '#/components/schemas/ReportClaim'
        createdAt:
          $ref: '#/components/schemas/UtcDateTime'
      additionalProperties: false
...
    CreateRcaRunRequest:
      type: object
      properties:
        reason:
          type: string
          minLength: 1
          maxLength: 2000
      additionalProperties: false
...
    CursorPageIncidents:
      type: object
      required: [items, nextCursor]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/IncidentSummary'
        nextCursor:
          type: [string, 'null']
      additionalProperties: false
...
    CursorPageRcaRuns:
      type: object
      required: [items, nextCursor]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/RcaRun'
        nextCursor:
          type: [string, 'null']
      additionalProperties: false
...